+++ Develop/db/db.json
@@ -22,5 +22,34 @@
     {
         "title": "cheers",
         "text": "hey"
+    },
+    {
+        "title": "Jacob",
+        "text": "Lost"
+    },
+    {
+        "title": "askd",
+        "text": "sldkanskd",
+        "noteId": "257a"
+    },
+    {
+        "title": "Cheers",
+        "text": "nkwdwa",
+        "noteId": "3c93"
+    },
+    {
+        "title": "please worl",
+        "text": "????",
+        "noteId": "356f"
+    },
+    {
+        "title": "onasfkjns",
+        "text": "zslkdalskdnas",
+        "noteId": "3a03"
+    },
+    {
+        "title": "wtf",
+        "text": "wtffff",
+        "noteId": "d4ce"
     }
 ]
