+++ Develop/server.js
@@ -1,11 +1,13 @@
 const express = require('express');
 const path = require('path');
 const fs = require('fs');
+// const getNotes = require('./public/assets/js/index')
 
 const PORT = 3001;
 const app = express();
+const api = require('./routes/index')
 
-const notes = require('./db/db.json')
+// const notes = require('./db/db.json')
 
 // middleware
 app.use(express.json());
@@ -22,59 +24,8 @@ app.get('/notes', (req, res) =>
     res.sendFile(path.join(__dirname, '/public/notes.html'))
 )
 
-app.get('/api/notes', (req, res) => {
-    // send db.json to client   
-    res.status(200).json(notes)
-
-    // Log our request to the terminal
-    console.info(`${req.method} request received to get notes`);
-
-
-})
-
-app.post('/api/notes', (req, res) => {
-    console.info(`${req.method} request received to add a review`);
-
-    const { title, text } = req.body;
-
-    if (title && text) {
-        const newNote = {
-            title,
-            text
-        }
-        // Read existing note database
-        fs.readFile(`./db/db.json`, 'utf-8', (err, data) => {
-
-            if (err) {
-                console.error(err);
-            } else {
-
-                // parse note data base and add new note
-                const noteArray = JSON.parse(data);
-                noteArray.push(newNote)
-
-                // write updated note object array to database
-                fs.writeFile('./db/db.json', JSON.stringify(noteArray, null, 4), (err) =>
-                    err ? console.error(err) : console.info(`\nData written to ./db/db.json`)
-                );
-
-                const response = {
-                    status: 'success',
-                    body: noteArray
-                };
-        
-                console.log(response);
-                res.status(201).json(response);
-            }
-
-        });
-        
-
-    } else {
-        res.status(500).json('Error in creating new note');
-    }
-
-})
+// adds api routes for note POST and GET
+app.use('/api', api)
 
 app.listen(PORT, () =>
     console.log(`App listening at http://localhost:${PORT} 🚀`)
